Fix signed/unsigned comparisons in tests

+++ tests/test_validation_errors.cpp
@@ -53,40 +53,40 @@ TEST_F(TestValidationErrors, AllOfConstraintFailure)
     ValidationResults::Error error;
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 2, error.context.size() );
+    EXPECT_EQ( size_t(2), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "[0]", error.context[1] );
     EXPECT_EQ( "Value type not permitted by 'type' constraint.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 1, error.context.size() );
+    EXPECT_EQ( size_t(1), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "Failed to validate item #0 in array.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 2, error.context.size() );
+    EXPECT_EQ( size_t(2), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "[1]", error.context[1] );
     EXPECT_EQ( "Value type not permitted by 'type' constraint.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 1, error.context.size() );
+    EXPECT_EQ( size_t(1), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "Failed to validate item #1 in array.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 2, error.context.size() );
+    EXPECT_EQ( size_t(2), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "[2]", error.context[1] );
     EXPECT_EQ( "Value type not permitted by 'type' constraint.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 1, error.context.size() );
+    EXPECT_EQ( size_t(1), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "Failed to validate item #2 in array.", error.description );
 
     EXPECT_TRUE( results.popError(error) );
-    EXPECT_EQ( 1, error.context.size() );
+    EXPECT_EQ( size_t(1), error.context.size() );
     EXPECT_EQ( "<root>", error.context[0] );
     EXPECT_EQ( "Failed to validate against child schema #0.", error.description );
